Создает отдельную папку для утилит, разделяет файл utils.js на common.js и render.js и корректирует импорты.
